embedding-host/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pillow==11.3.0
torch>=2.4.0
transformers>=4.57.0
accelerate>=1.0.0
safetensors>=0.4.0
qwen-vl-utils>=0.0.14
torchvision
torchaudio